lib/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
boto3==1.40.59
botocore==1.40.59
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jmespath==1.0.1
lxml==6.0.2
oauthlib==3.3.1
openai==2.6.1
outcome==1.3.0.post0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.14.0
selenium==4.36.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
tweepy==4.16.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3>=1.26.20,<3
webencodings==0.5.1
websocket-client==1.9.0
wsproto==1.2.0
